ui: Refactor Metrics components
Refactor of metrics-related components/containers, which was the worst
offender in terms of unecessary circular dependencies arising from rapid
development.

+ "views/shared/components/metricQuery" is now the bottom of the stack,
containing all of the "query selector" components and interfaces.
+ Created a new "views/cluster/util/graphs.ts" to contain graph-related
utility functions. These functions do not produce React elements, but are used to perform valuable calculations related to rendering graphs.
+ Refactored metricDataProvider, graphs.ts, linegraph, and summaryBar to align with the new factoring.
